Remove older AP driver and remove experimental name from it
