+++ drivers.xml
@@ -49,18 +49,14 @@
             <driver name="EQ500X">indi_eq500x_telescope</driver>
             <version>1.0</version>
         </device>
-        <device label="AstroPhysics Experimental" manufacturer="Astrophysics">
-            <driver name="AstroPhysics Experimental">indi_lx200ap_experimental</driver>
+        <device label="AstroPhysics" manufacturer="Astrophysics">
+            <driver name="AstroPhysics">indi_lx200ap</driver>
             <version>1.0</version>
         </device>
         <device label="AstroPhysics GTOCP2" manufacturer="Astrophysics">
             <driver name="AstroPhysics GTOCP2">indi_lx200ap_gtocp2</driver>
             <version>1.0</version>
-        </device>
-        <device label="AstroPhysics" manufacturer="Astrophysics">
-            <driver name="AstroPhysics">indi_lx200ap</driver>
-            <version>1.0</version>
-        </device>
+        </device>        
         <device label="Celestron GPS" manufacturer="Celestron">
             <driver name="Celestron GPS">indi_celestron_gps</driver>
             <version>3.6</version>
