--- drivers/telescope/lx200ap_experimentaldriver.h
@@ -1,35 +1,0 @@
-/*
-    LX200 AP Driver
-    Copyright (C) 2007 Markus Wildi
-
-    This library is free software; you can redistribute it and/or
-    modify it under the terms of the GNU Lesser General Public
-    License as published by the Free Software Foundation; either
-    version 2.1 of the License, or (at your option) any later version.
-
-    This library is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-    Lesser General Public License for more details.
-
-    You should have received a copy of the GNU Lesser General Public
-    License along with this library; if not, write to the Free Software
-    Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
-
-*/
-
-#pragma once
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-void set_lx200ap_exp_name(const char *deviceName, unsigned int debug_level);
-int selectAPCenterRate(int fd, int centerRate);
-int check_lx200ap_status(int fd, char *parkStatus, char *slewStatus);
-int APParkMount(int fd);
-int APUnParkMount(int fd);
-
-#ifdef __cplusplus
-}
-#endif
