+++ drivers/telescope/lx200apdriver.cpp
@@ -10,7 +10,7 @@ either
 version 2.1 of the License, or (at your option) any later version.
 
 This library is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY;
+     but WITHOUT ANY WARRANTY;
 without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 Lesser General Public License for more details.
@@ -135,31 +135,31 @@ int getAPUTCOffset(int fd, double *value)
             temp_string[1] = '0';
             switch (temp_string[2])
             {
-            case '5':
+                case '5':
 
-                temp_string[2] = '1';
-                break;
-            case '4':
+                    temp_string[2] = '1';
+                    break;
+                case '4':
 
-                temp_string[2] = '2';
-                break;
-            case '3':
+                    temp_string[2] = '2';
+                    break;
+                case '3':
 
-                temp_string[2] = '3';
-                break;
-            case '2':
+                    temp_string[2] = '3';
+                    break;
+                case '2':
 
-                temp_string[2] = '4';
-                break;
-            case '1':
+                    temp_string[2] = '4';
+                    break;
+                case '1':
 
-                temp_string[2] = '5';
-                break;
-            default:
-                DEBUGFDEVICE(lx200ap_name, INDI::Logger::DBG_ERROR, "getAPUTCOffset: string not handled %s",
-                             temp_string);
-                return -1;
-                break;
+                    temp_string[2] = '5';
+                    break;
+                default:
+                    DEBUGFDEVICE(lx200ap_name, INDI::Logger::DBG_ERROR, "getAPUTCOffset: string not handled %s",
+                                 temp_string);
+                    return -1;
+                    break;
             }
         }
         else if (temp_string[1] == '0')
@@ -172,36 +172,36 @@ int getAPUTCOffset(int fd, double *value)
             temp_string[1] = '0';
             switch (temp_string[2])
             {
-            case '9':
+                case '9':
 
-                temp_string[2] = '7';
-                break;
-            case '8':
+                    temp_string[2] = '7';
+                    break;
+                case '8':
 
-                temp_string[2] = '8';
-                break;
-            case '7':
+                    temp_string[2] = '8';
+                    break;
+                case '7':
 
-                temp_string[2] = '9';
-                break;
-            case '6':
+                    temp_string[2] = '9';
+                    break;
+                case '6':
 
-                temp_string[2] = '0';
-                break;
-            case '5':
-                temp_string[1] = '1';
-                temp_string[2] = '1';
-                break;
-            case '4':
+                    temp_string[2] = '0';
+                    break;
+                case '5':
+                    temp_string[1] = '1';
+                    temp_string[2] = '1';
+                    break;
+                case '4':
 
-                temp_string[1] = '1';
-                temp_string[2] = '2';
-                break;
-            default:
-                DEBUGFDEVICE(lx200ap_name, INDI::Logger::DBG_ERROR, "getAPUTCOffset: string not handled %s",
-                             temp_string);
-                return -1;
-                break;
+                    temp_string[1] = '1';
+                    temp_string[2] = '2';
+                    break;
+                default:
+                    DEBUGFDEVICE(lx200ap_name, INDI::Logger::DBG_ERROR, "getAPUTCOffset: string not handled %s",
+                                 temp_string);
+                    return -1;
+                    break;
             }
         }
         else
@@ -335,29 +335,29 @@ int selectAPPECState(int fd, int pecstate)
 
     switch (pecstate)
     {
-    // PEC OFF
-    case 0:
-        DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPPECState: Setting PEC OFF");
-        DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:p#");
+        // PEC OFF
+        case 0:
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPPECState: Setting PEC OFF");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:p#");
 
-        if ((error_type = tty_write_string(fd, "#:p#", &nbytes_write)) != TTY_OK)
-            return error_type;
+            if ((error_type = tty_write_string(fd, "#:p#", &nbytes_write)) != TTY_OK)
+                return error_type;
 
-        break;
+            break;
 
         // PEC ON
-    case 1:
-        DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPPECState: Setting PEC ON");
-        DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:pP#");
+        case 1:
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPPECState: Setting PEC ON");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:pP#");
 
-        if ((error_type = tty_write_string(fd, "#:pP#", &nbytes_write)) != TTY_OK)
-            return error_type;
+            if ((error_type = tty_write_string(fd, "#:pP#", &nbytes_write)) != TTY_OK)
+                return error_type;
 
-        break;
+            break;
 
-    default:
-        return -1;
-        break;
+        default:
+            return -1;
+            break;
     }
 
     return 0;
@@ -370,44 +370,44 @@ int selectAPMoveToRate(int fd, int moveToRate)
 
     switch (moveToRate)
     {
-    /* 12x*/
-    case 0:
-        DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPMoveToRate: Setting move to rate to 12x");
-        DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RC0#");
+        /* 12x*/
+        case 0:
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPMoveToRate: Setting move to rate to 12x");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RC0#");
 
-        if ((error_type = tty_write_string(fd, "#:RC0#", &nbytes_write)) != TTY_OK)
-            return error_type;
-        break;
+            if ((error_type = tty_write_string(fd, "#:RC0#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
 
-    /* 64x */
-    case 1:
-        DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPMoveToRate: Setting move to rate to 64x");
-        DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RC1#");
+        /* 64x */
+        case 1:
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPMoveToRate: Setting move to rate to 64x");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RC1#");
 
-        if ((error_type = tty_write_string(fd, "#:RC1#", &nbytes_write)) != TTY_OK)
-            return error_type;
-        break;
+            if ((error_type = tty_write_string(fd, "#:RC1#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
 
-    /* 600x */
-    case 2:
-        DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPMoveToRate: Setting move to rate to 600x");
-        DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RC2#");
-        if ((error_type = tty_write_string(fd, "#:RC2#", &nbytes_write)) != TTY_OK)
-            return error_type;
-        break;
+        /* 600x */
+        case 2:
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPMoveToRate: Setting move to rate to 600x");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RC2#");
+            if ((error_type = tty_write_string(fd, "#:RC2#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
 
-    /* 1200x */
-    case 3:
-        DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPMoveToRate: Setting move to rate to 1200x");
-        DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RC3#");
+        /* 1200x */
+        case 3:
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPMoveToRate: Setting move to rate to 1200x");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RC3#");
 
-        if ((error_type = tty_write_string(fd, "#:RC3#", &nbytes_write)) != TTY_OK)
-            return error_type;
-        break;
+            if ((error_type = tty_write_string(fd, "#:RC3#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
 
-    default:
-        return -1;
-        break;
+        default:
+            return -1;
+            break;
     }
     return 0;
 }
@@ -418,40 +418,40 @@ int selectAPSlewRate(int fd, int slewRate)
     int nbytes_write = 0;
     switch (slewRate)
     {
-    /* 600x */
-    case 0:
+        /* 600x */
+        case 0:
 
-        DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPSlewRate: Setting slew to rate to 600x");
-        DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RS0#");
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPSlewRate: Setting slew to rate to 600x");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RS0#");
 
-        if ((error_type = tty_write_string(fd, "#:RS0#", &nbytes_write)) != TTY_OK)
-            return error_type;
-        break;
+            if ((error_type = tty_write_string(fd, "#:RS0#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
 
-    /* 900x */
-    case 1:
+        /* 900x */
+        case 1:
 
-        DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPSlewRate: Setting slew to rate to 900x");
-        DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RS1#");
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPSlewRate: Setting slew to rate to 900x");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RS1#");
 
-        if ((error_type = tty_write_string(fd, "#:RS1#", &nbytes_write)) != TTY_OK)
-            return error_type;
-        break;
+            if ((error_type = tty_write_string(fd, "#:RS1#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
 
 
-    /* 1200x */
-    case 2:
+        /* 1200x */
+        case 2:
 
-        DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPSlewRate: Setting slew to rate to 1200x");
-        DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RS2#");
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPSlewRate: Setting slew to rate to 1200x");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RS2#");
 
-        if ((error_type = tty_write_string(fd, "#:RS2#", &nbytes_write)) != TTY_OK)
-            return error_type;
-        break;
+            if ((error_type = tty_write_string(fd, "#:RS2#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
 
-    default:
-        return -1;
-        break;
+        default:
+            return -1;
+            break;
     }
     return 0;
 }
@@ -463,54 +463,54 @@ int selectAPTrackingMode(int fd, int trackMode)
 
     switch (trackMode)
     {
-    /* Sidereal */
-    case AP_TRACKING_SIDEREAL:
+        /* Sidereal */
+        case AP_TRACKING_SIDEREAL:
 
-        DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG,
-                    "selectAPTrackingMode: Setting tracking mode to sidereal.");
-        DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RT2#");
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG,
+                        "selectAPTrackingMode: Setting tracking mode to sidereal.");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RT2#");
 
-        if ((error_type = tty_write_string(fd, "#:RT2#", &nbytes_write)) != TTY_OK)
-            return error_type;
-        break;
+            if ((error_type = tty_write_string(fd, "#:RT2#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
 
         /* Solar */
-    case AP_TRACKING_SOLAR:
+        case AP_TRACKING_SOLAR:
 
-        DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPTrackingMode: Setting tracking mode to solar.");
-        DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RT1#");
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPTrackingMode: Setting tracking mode to solar.");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RT1#");
 
-        if ((error_type = tty_write_string(fd, "#:RT1#", &nbytes_write)) != TTY_OK)
-            return error_type;
-        break;
+            if ((error_type = tty_write_string(fd, "#:RT1#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
 
         /* Lunar */
-    case AP_TRACKING_LUNAR:
+        case AP_TRACKING_LUNAR:
 
-        DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPTrackingMode: Setting tracking mode to lunar.");
-        DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RT0#");
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPTrackingMode: Setting tracking mode to lunar.");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RT0#");
 
-        if ((error_type = tty_write_string(fd, "#:RT0#", &nbytes_write)) != TTY_OK)
-            return error_type;
-        break;
+            if ((error_type = tty_write_string(fd, "#:RT0#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
 
-    case AP_TRACKING_CUSTOM:
+        case AP_TRACKING_CUSTOM:
             DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPTrackingMode: Setting tracking mode to Custom.");
             break;
 
         /* Zero */
-    case AP_TRACKING_OFF:
+        case AP_TRACKING_OFF:
 
-        DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPTrackingMode: Setting tracking mode to Zero.");
-        DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RT9#");
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPTrackingMode: Setting tracking mode to Zero.");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RT9#");
 
-        if ((error_type = tty_write_string(fd, "#:RT9#", &nbytes_write)) != TTY_OK)
-            return error_type;
-        break;
+            if ((error_type = tty_write_string(fd, "#:RT9#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
 
-    default:
-        return -1;
-        break;
+        default:
+            return -1;
+            break;
     }
     return 0;
 }
@@ -521,40 +521,40 @@ int selectAPGuideRate(int fd, int guideRate)
     int nbytes_write = 0;
     switch (guideRate)
     {
-    /* 0.25x */
-    case 0:
+        /* 0.25x */
+        case 0:
 
-        DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPGuideRate: Setting guide to rate to 0.25x");
-        DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RG0#");
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPGuideRate: Setting guide to rate to 0.25x");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RG0#");
 
-        if ((error_type = tty_write_string(fd, "#:RG0#", &nbytes_write)) != TTY_OK)
-            return error_type;
-        break;
+            if ((error_type = tty_write_string(fd, "#:RG0#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
 
-    /* 0.50x */
-    case 1:
+        /* 0.50x */
+        case 1:
 
-        DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPGuideRate: Setting guide to rate to 0.50x");
-        DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RG1#");
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPGuideRate: Setting guide to rate to 0.50x");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RG1#");
 
-        if ((error_type = tty_write_string(fd, "#:RG1#", &nbytes_write)) != TTY_OK)
-            return error_type;
-        break;
+            if ((error_type = tty_write_string(fd, "#:RG1#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
 
 
-    /* 1.00x */
-    case 2:
+        /* 1.00x */
+        case 2:
 
-        DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPGuideRate: Setting guide to rate to 1.00x");
-        DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RG2#");
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPGuideRate: Setting guide to rate to 1.00x");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RG2#");
 
-        if ((error_type = tty_write_string(fd, "#:RG2#", &nbytes_write)) != TTY_OK)
-            return error_type;
-        break;
+            if ((error_type = tty_write_string(fd, "#:RG2#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
 
-    default:
-        return -1;
-        break;
+        default:
+            return -1;
+            break;
     }
     return 0;
 }
@@ -566,22 +566,22 @@ int swapAPButtons(int fd, int currentSwap)
 
     switch (currentSwap)
     {
-    case 0:
+        case 0:
 
-        DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:NS#");
-        if ((error_type = tty_write_string(fd, "#:NS#", &nbytes_write)) != TTY_OK)
-            return error_type;
-        break;
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:NS#");
+            if ((error_type = tty_write_string(fd, "#:NS#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
 
-    case 1:
-        DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:EW#");
-        if ((error_type = tty_write_string(fd, "#:EW#", &nbytes_write)) != TTY_OK)
-            return error_type;
-        break;
+        case 1:
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:EW#");
+            if ((error_type = tty_write_string(fd, "#:EW#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
 
-    default:
-        return -1;
-        break;
+        default:
+            return -1;
+            break;
     }
     return 0;
 }
@@ -765,13 +765,13 @@ int APSendPulseCmd(int fd, int direction, int duration_msec)
             break;
         case LX200_SOUTH:
             sprintf(cmd, ":Ms%03d#", duration_msec);
-        break;
+            break;
         case LX200_EAST:
             sprintf(cmd, ":Me%03d#", duration_msec);
-        break;
+            break;
         case LX200_WEST:
             sprintf(cmd, ":Mw%03d#", duration_msec);
-        break;
+            break;
         default:
             return 1;
     }
@@ -831,3 +831,143 @@ int check_lx200ap_status(int fd, char *parkStatus, char *slewStatus)
     return -1;
 }
 #endif
+
+// make this a function with logging instead of a #define like in legacy driver
+int APParkMount(int fd)
+{
+    int error_type;
+    int nbytes_write = 0;
+
+    DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "APParkMount: Sending park command.");
+    DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:KA");
+
+    if ((error_type = tty_write_string(fd, "#:KA", &nbytes_write)) != TTY_OK)
+        return error_type;
+
+    return 0;
+}
+
+// make this a function with logging instead of a #define like in legacy driver
+int APUnParkMount(int fd)
+{
+    int error_type;
+    int nbytes_write = 0;
+
+    DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "APUnParkMount: Sending unpark command.");
+    DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:PO");
+
+    if ((error_type = tty_write_string(fd, "#:PO", &nbytes_write)) != TTY_OK)
+        return error_type;
+
+    return 0;
+}
+
+// This is a modified version of selectAPMoveRate() from lx200apdriver.cpp
+// This version allows changing the rate to GUIDE as well as 12x/64x/600x/1200x
+// and is required some the experimental AP driver properly handles
+// pulse guide requests over 999ms by simulated it by setting the move rate
+// to GUIDE and then starting and halting a move of the correct duration.
+int selectAPCenterRate(int fd, int centerRate)
+{
+    int error_type;
+    int nbytes_write = 0;
+
+    switch (centerRate)
+    {
+        /* Guide */
+        case 0:
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPMoveToRate: Setting move to rate to GUIDE");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RG#");
+
+            if ((error_type = tty_write_string(fd, "#:RG#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
+
+        /* 12x */
+        case 1:
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPMoveToRate: Setting move to rate to 12x");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RC0#");
+
+            if ((error_type = tty_write_string(fd, "#:RC0#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
+
+        /* 64x */
+        case 2:
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPMoveToRate: Setting move to rate to 64x");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RC1#");
+
+            if ((error_type = tty_write_string(fd, "#:RC1#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
+
+        /* 600x */
+        case 3:
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPMoveToRate: Setting move to rate to 600x");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RC2#");
+            if ((error_type = tty_write_string(fd, "#:RC2#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
+
+        /* 1200x */
+        case 4:
+            DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "selectAPMoveToRate: Setting move to rate to 1200x");
+            DEBUGFDEVICE(lx200ap_name, AP_DBG_SCOPE, "CMD <%s>", "#:RC3#");
+
+            if ((error_type = tty_write_string(fd, "#:RC3#", &nbytes_write)) != TTY_OK)
+                return error_type;
+            break;
+
+        default:
+            return -1;
+            break;
+    }
+    return 0;
+}
+
+// experimental functions!!!
+
+
+int check_lx200ap_status(int fd, char *parkStatus, char *slewStatus)
+{
+    char temp_string[64];
+    int error_type;
+    int nbytes_write = 0;
+    int nbytes_read  = 0;
+
+    DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "Check status...");
+
+    if (fd <= 0)
+    {
+        DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_ERROR,
+                    "check_lx200ap_connection: not a valid file descriptor received");
+
+        return -1;
+    }
+
+    if ((error_type = tty_write_string(fd, "#:GOS#", &nbytes_write)) != TTY_OK)
+    {
+        DEBUGFDEVICE(lx200ap_name, INDI::Logger::DBG_ERROR,
+                     "check_lx200ap_connection: unsuccessful write to telescope, %d", nbytes_write);
+
+        return error_type;
+    }
+    tty_read_section(fd, temp_string, '#', LX200_TIMEOUT, &nbytes_read);
+    tcflush(fd, TCIFLUSH);
+    if (nbytes_read > 1)
+    {
+        temp_string[nbytes_read - 1] = '\0';
+
+        DEBUGFDEVICE(lx200ap_name, INDI::Logger::DBG_DEBUG, "check_lx200ap_status: received bytes %d, [%s]",
+                     nbytes_write, temp_string);
+
+        *parkStatus = temp_string[0];
+        *slewStatus = temp_string[3];
+
+        return 0;
+    }
+
+    DEBUGDEVICE(lx200ap_name, INDI::Logger::DBG_ERROR, "check_lx200ap_status: wrote, but nothing received.");
+
+    return -1;
+}
